relax the language reporting details checks a bit, and update policy checklist methods as per #367
