+++ process/subprocesses/ghsci.py
@@ -195,130 +195,6 @@ def get_analysis_report_region_configuration(region_config, settings):
     return region_config
 
 
-def check_and_update_config_reporting_parameters(config):
-    from _utils import (
-        get_valid_languages,
-        print_autobreak,
-        setup_default_language,
-    )
-
-    """Checks config reporting parameters and updates these if necessary."""
-    reporting_default = {
-        'configuration': './configuration/_report_configuration.xlsx',
-        'templates': ['policy_spatial'],
-        'publication_ready': False,
-        'doi': None,
-        'images': {
-            1: {
-                'file': 'Example image of a vibrant, walkable, urban neighbourhood - landscape.jpg',
-                'description': 'Example image of a vibrant, walkable, urban neighbourhood with diverse people using active modes of transport and a tram (replace with a photograph, customised in region configuration)',
-                'credit': 'Carl Higgs, Bing Image Creator, 2023',
-            },
-            2: {
-                'file': 'Example image of a vibrant, walkable, urban neighbourhood - square.jpg',
-                'description': 'Example image of a vibrant, walkable, urban neighbourhood with diverse people using active modes of transport and a tram (replace with a photograph, customised in region configuration)',
-                'credit': 'Carl Higgs, Bing Image Creator, 2023',
-            },
-        },
-        'languages': setup_default_language(config),
-        'exceptions': {},
-    }
-    if 'reporting' not in config:
-        print_autobreak(
-            "\nNote: No 'reporting' section found in region configuration.  This is required for report generation.  A default parameterisation will be used for reporting in English.  To further customise for your region and requirements, please add and update the reporting section in your region's configuration file.",
-        )
-        reporting = reporting_default.copy()
-    else:
-        reporting = config['reporting'].copy()
-    for key in reporting_default.keys():
-        if key not in reporting.keys():
-            reporting[key] = reporting_default[key]
-            print_autobreak(
-                f"\nNote: Reporting parameter '{key}' not found in region configuration.  Using default value of '{reporting_default[key]}'.  To further customise for your region and requirements, please add and update the reporting section in your region's configuration file.",
-            )
-    config['reporting'] = reporting
-    reporting['languages'] = get_valid_languages(config)
-    return reporting
-
-
-def get_valid_languages(config):
-    """Check if language is valid for given configuration."""
-    from _utils import print_autobreak, setup_default_language
-
-    no_language_warning = "No valid languages found in region configuration.  This is required for report generation.  A default parameterisation will be used for reporting in English.  To further customise for your region and requirements, please add and update the reporting section in your region's configuration file."
-    default_language = setup_default_language(config)
-    configured_languages = pd.read_excel(
-        config['reporting']['configuration'], sheet_name='languages',
-    ).columns[2:]
-    configured_fonts = pd.read_excel(
-        config['reporting']['configuration'], sheet_name='fonts',
-    )
-    if config['reporting']['languages'] is None:
-        print_autobreak(f'\nNote: {no_language_warning}')
-        languages = default_language
-    else:
-        languages = config['reporting']['languages']
-
-    languages_configured = [x for x in languages if x in configured_languages]
-    if len(languages_configured) == 0:
-        print_autobreak(f'\nNote: {no_language_warning}')
-        languages = default_language
-    else:
-        if len(languages_configured) < len(languages):
-            languages_not_configured = [
-                x for x in languages if x not in configured_languages
-            ]
-            print_autobreak(
-                f"\nNote: Some languages specified in this region's configuration file ({', '.join(languages_not_configured)}) have not been set up with translations in the report configuration 'languages' worksheet.  Reports will only be generated for those languages that have had prose translations set up ({', '.join(configured_languages)}).",
-            )
-    required_keys = {'country', 'summary', 'name', 'context'}
-    languages_configured_have_required_keys = [
-        x for x in languages_configured if languages[x].keys() == required_keys
-    ]
-    if len(languages_configured_have_required_keys) < len(
-        languages_configured,
-    ):
-        languages_configured_without_required_keys = [
-            x
-            for x in languages_configured
-            if x not in languages_configured_have_required_keys
-        ]
-        missing_keys = {
-            m: [x for x in required_keys if x not in languages[m].keys()]
-            for m in languages_configured_without_required_keys
-        }
-        print_autobreak(
-            f"""\nNote: Some configured languages ({languages_configured_without_required_keys}) do not have all the required keys (missing or mis-spelt keys: {missing_keys}).  These will be set up to use default values.""",
-        )
-        for language in languages_configured_without_required_keys:
-            for key in required_keys:
-                if key not in languages[language].keys():
-                    languages[language][key] = default_language['English'][key]
-    languages = {
-        cl: languages[cl] for cl in languages if cl in configured_languages
-    }
-    for font_language in set(
-        configured_fonts.loc[configured_fonts['Language'].isin(languages)][
-            'Language'
-        ],
-    ):
-        language_fonts_list = (
-            configured_fonts.loc[
-                configured_fonts['Language'] == font_language
-            ]['File']
-            .unique()
-            .tolist()
-        )
-        if not all([os.path.exists(x) for x in language_fonts_list]):
-            print_autobreak(
-                f"\nNote: One or more fonts specified in this region's configuration file for the language {font_language} ({', '.join(language_fonts_list)}) do not exist.  This language will be skipped when generating maps, figures and reports until configured fonts can be located.  These may have to be downloaded and stored in the configured location.",
-            )
-            languages = {
-                f: languages[f] for f in languages if f != font_language
-            }
-    return languages
-
-
 def format_date(date, format='%Y-%m-%d'):
     """Format date as string."""
     from datetime import date as datetime_date
@@ -467,7 +343,6 @@ class Region:
                 'policy_review'
             ] = f'{folder_path}/process/data/policy_review/_policy_review_template_v0_TO-BE-UPDATED.xlsx'
         r = get_analysis_report_region_configuration(r, settings)
-        r['reporting'] = check_and_update_config_reporting_parameters(r)
         return r
 
     def _verify_data_dir(self, data_dir, verify_file_extension=None) -> dict:
@@ -670,10 +545,16 @@ class Region:
 
     def generate_report(self, language: str = 'English', report='indicators'):
         """Generate a report for this study region."""
-        from _utils import generate_report_for_language
+        from _utils import (
+            check_and_update_config_reporting_parameters,
+            generate_report_for_language,
+        )
         from subprocesses.analysis_report import PDF_Analysis_Report
 
         if report == 'indicators':
+            self.config[
+                'reporting'
+            ] = check_and_update_config_reporting_parameters(self.config)
             generate_report_for_language(
                 self, language, indicators, policies,
             )
